chore: TicketList -> Ticket으로 이름 변경, pages 업데이트
