+++ src/components/common/Router.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import Main from '../../pages/Main';
 import PriceCompare from '../../pages/PriceCompare';
-import TicketList from '../ticketList/TicketList';
+import TicketList from '../../pages/TicketList';
 const Router = () => {
   return (
     <BrowserRouter>
