--- src/pages/TickeList.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const TickeList = () => {
-  return (
-    <div>TickeList</div>
-  )
-}
-
-export default TickeList
